@@ -31,7 +31,6 @@
     "babel-polyfill": "^6.26.0",
     "core-js": "^3.1.4",
     "eventemitter2": "^5.0.1",
-    "formiojs": "^4.9.0",
     "lodash": "^4.17.15",
     "prop-types": "^15.7.2"
   },
@@ -57,6 +56,7 @@
     "eslint-plugin-jsx-a11y": "^2.0.1",
     "eslint-plugin-mocha": "^5.2.0",
     "eslint-plugin-react": "^6.0.0",
+    "formiojs": "^4.9.0",
     "jsdom": "^9.4.2",
     "karma": "^4.4.1",
     "karma-chrome-launcher": "^2.2.0",
@@ -64,9 +64,9 @@
     "karma-mocha-reporter": "^2.2.5",
     "karma-webpack": "^3.0.5",
     "mocha": "^5.2.0",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-test-renderer": "^16.5.2",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-test-renderer": "^16.13.1",
     "react-transform-catch-errors": "^1.0.0",
     "sinon": "^7.0.0",
     "webpack": "^2.7.0"
@@ -75,6 +75,7 @@
     "lib"
   ],
   "peerDependencies": {
+    "formiojs": "^4.9.0",
     "react": "^15.3.0 || ^16.0.0",
     "react-dom": "^15.3.0 || ^16.0.0"
   }
