@@ -73,17 +73,16 @@ function Pagination({
     <nav aria-label="Page navigation">
       <ul className="pagination">
         <li className={`page-item ${(activePage === 1) ? 'disabled' : ''}`}>
-          <a
+          <button
             className="page-link"
             onClick={() => {
               if (activePage !== 1) {
                 onSelect(activePage - 1);
               }
             }}
-            href="javascript:void(0)"
           >
             {prev}
-          </a>
+          </button>
         </li>
 
         {
@@ -102,30 +101,28 @@ function Pagination({
 
             return (
               <li className={`page-item ${className}`} key={page}>
-                <a
+                <button
                   className="page-link"
                   onClick={() => onSelect(page)}
-                  href="javascript:void(0)"
                 >
                   {page}
-                </a>
+                </button>
               </li>
             );
           })
         }
 
         <li className={`page-item ${(activePage === pages) ? 'disabled' : ''}`}>
-          <a
+          <button
             className="page-link"
             onClick={() => {
               if (activePage !== pages) {
                 onSelect(activePage + 1);
               }
             }}
-            href="javascript:void(0)"
           >
             {next}
-          </a>
+          </button>
         </li>
       </ul>
     </nav>
